+++ cs/cs/Serializer/VowpalWabbitJsonSerializer.cs
@@ -26,10 +26,40 @@ namespace VW.Serializer
     {
         internal const string FeatureIgnorePrefix = "_";
 
+        public interface IResult
+        {
+        }
+
+        public sealed class SingleResult : IResult
+        {
+            /// <summary>
+            /// Single example or the shared
+            /// </summary>
+            public VowpalWabbitExample Example { get; set; }
+        }
+
+        public sealed class MultilineResult : IResult
+        {
+            /// <summary>
+            /// Single example or the shared
+            /// </summary>
+            public VowpalWabbitExample Example { get; set; }
+
+            /// <summary>
+            /// The multi-line examples
+            /// </summary>
+            public List<VowpalWabbitExample> Examples { get; set; }
+        }
+
+
         /// <summary>
         /// Mapping from properties to types for labels.
         /// </summary>
         private static readonly Dictionary<string, Type> labelPropertyMapping;
+
+        private readonly VowpalWabbit vw;
+        private readonly VowpalWabbitDefaultMarshaller defaultMarshaller;
+        private readonly JsonSerializer jsonSerializer;
 
         static VowpalWabbitJsonSerializer()
         {
@@ -47,10 +77,6 @@ namespace VW.Serializer
 
             labelPropertyMapping = q.ToDictionary(e => e.JsonProperty.PropertyName ?? e.Property.Name, e => e.Type);
         }
-        private readonly VowpalWabbit vw;
-        private readonly VowpalWabbitDefaultMarshaller defaultMarshaller;
-        private readonly JsonSerializer jsonSerializer;
-        private JsonReader reader;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="VowpalWabbitJson"/> class.
@@ -64,19 +90,18 @@ namespace VW.Serializer
             this.defaultMarshaller = new VowpalWabbitDefaultMarshaller();
             this.jsonSerializer = new JsonSerializer();
 
-            this.Context = new VowpalWabbitMarshalContext(this.vw);
-            this.DefaultNamespaceContext = new VowpalWabbitMarshalContext(this.vw, this.Context.ExampleBuilder);
+            this.ExampleBuilder = new VowpalWabbitJsonBuilder(vw, this.defaultMarshaller, this.jsonSerializer);
         }
 
         /// <summary>
-        /// The marshalling context. Can be modified until <see cref="CreateExample"/>.
+        /// Single line example or shared example.
         /// </summary>
-        public VowpalWabbitMarshalContext Context { get; private set; }
+        public VowpalWabbitJsonBuilder ExampleBuilder { get; private set; }
 
         /// <summary>
-        /// The marshalling context for the default namespace. Can be modified until <see cref="CreateExample"/>.
+        /// Multi-line examples.
         /// </summary>
-        public VowpalWabbitMarshalContext DefaultNamespaceContext { get; private set; }
+        public List<VowpalWabbitJsonBuilder> ExampleBuilders { get; private set; }
 
         /// <summary>
         /// Parses the example.
@@ -86,7 +111,7 @@ namespace VW.Serializer
         /// Optional label, taking precedence over "_label" property found in <paramref name="json"/>.
         /// If null, <paramref name="json"/> will be inspected and the "_label" property used as label.
         /// </param>
-        public void Parse(string json, ILabel label = null)
+        public void Parse(string json, ILabel label = null, int? index = null)
         {
             using (var textReader = new JsonTextReader(new StringReader(json)))
             {
@@ -94,109 +119,53 @@ namespace VW.Serializer
             }
         }
 
-        /// <summary>
-        /// Parses the example.
-        /// </summary>
-        /// <param name="reader">The example to parse.</param>
-        /// <param name="label">
-        /// Optional label, taking precedence over "_label" property found in <paramref name="reader"/>.
-        /// If null, <paramref name="reader"/> will be inspected and the "_label" property used as label.
-        /// </param>
-        /// <returns>The VowpalWabbit native example.</returns>
-        public void Parse(JsonReader reader, ILabel label = null)
+        public void Parse(JsonReader reader, ILabel label = null, int? index = null)
         {
-            // avoid parameter passing for the sake of non-reentrantness
-            this.reader = reader;
-
-            if (label != null)
-                this.defaultMarshaller.MarshalLabel(this.Context, label);
-
-            // handle the case when the reader is already positioned at JsonToken.StartObject
-            if (reader.TokenType == JsonToken.None && !reader.Read())
-                return;
-
-            if (reader.TokenType != JsonToken.StartObject)
-                throw new VowpalWabbitJsonException(reader.Path, "Expected start object");
-
-            Namespace defaultNamespace = new Namespace(this.vw);
-            using (this.DefaultNamespaceContext.NamespaceBuilder = this.DefaultNamespaceContext.ExampleBuilder.AddNamespace(VowpalWabbitConstants.DefaultNamespace))
-            {
-                while (reader.Read())
+            // only pass the label if it's not targeted at a particular index
+            this.ExampleBuilder.Parse(reader, index == null ? label : null,
+                propertyName =>
                 {
-                    switch (reader.TokenType)
+                    if (propertyName != "_multi")
+                        return false;
+
+                    if (!reader.Read() || reader.TokenType != JsonToken.StartArray)
+                        throw new VowpalWabbitJsonException(reader.Path, "Expected start array for _multi");
+
+                    if (this.ExampleBuilders == null)
+                        this.ExampleBuilders = new List<VowpalWabbitJsonBuilder>();
+
+                    while (reader.Read())
                     {
-                        case JsonToken.PropertyName:
-                            var propertyName = (string)reader.Value;
-                            if (propertyName.StartsWith(FeatureIgnorePrefix))
-                            {
-                                // special fields
-                                switch (propertyName)
+                        switch(reader.TokenType)
+                        {
+                            case JsonToken.StartObject:
+                                VowpalWabbitJsonBuilder builder = null;
+
+                                try
                                 {
-                                    case "_label":
-                                        // passed in label has precedence
-                                        if (label == null)
-                                            this.ParseLabel();
-                                        else
-                                            reader.Skip();
-                                        break;
-                                    //case "_shared":
-                                    //    break;
-                                    //case "_multi":
-                                    //    break;
-                                    default:
-                                        reader.Skip();
-                                        break;
+                                    builder = new VowpalWabbitJsonBuilder(this.vw, this.defaultMarshaller, this.jsonSerializer);
+                                    this.ExampleBuilders.Add(builder);
                                 }
-                            }
-                            else
-                            {
-                                if (!reader.Read())
-                                    throw new VowpalWabbitJsonException(reader.Path, "Unexpected end");
+                                catch (Exception)
+                                {
+                                    builder.Dispose();
+                                    throw;
+                                }
 
-                                if (reader.TokenType == JsonToken.StartObject)
-                                    this.ParseNamespaceAndFeatures(this.Context, propertyName);
-                                else
-                                    this.ParseFeature(this.DefaultNamespaceContext, defaultNamespace, propertyName);
-
-                            }
-                            break;
-                        case JsonToken.EndObject:
-                            goto done;
+                                // pass the label to the selected example
+                                builder.Parse(reader, index != null && index == this.ExampleBuilders.Count - 1 ? label : null);
+                                break;
+                            case JsonToken.EndArray:
+                                return true;
+                            default:
+                                throw new VowpalWabbitJsonException(reader.Path, "Unexpected token: " + reader.TokenType);
+                        }
                     }
-                }
-            done:
+                });
 
-                // append default namespaces features if we found some
-                if (this.DefaultNamespaceContext.StringExample != null && this.DefaultNamespaceContext.StringExample.Length > 0)
-                {
-                    this.Context.StringExample.AppendFormat(CultureInfo.InvariantCulture,
-                        " | {0}", this.DefaultNamespaceContext.StringExample);
-                }
-            }
-        }
-
-        /// <summary>
-        /// Creates the managed example representation.
-        /// </summary>
-        /// <returns>Returns the managed example.</returns>
-        public VowpalWabbitExample CreateExample()
-        {
-            try
-	        {
-                var vwExample = this.Context.ExampleBuilder.CreateExample();
-
-                if (this.vw.Settings.EnableStringExampleGeneration)
-                    vwExample.VowpalWabbitString = this.Context.StringExample.ToString();
-
-                return vwExample;
-	        }
-	        finally
-	        {
-                this.DefaultNamespaceContext.Dispose();
-                this.DefaultNamespaceContext = null;
-                this.Context.Dispose();
-                this.Context = null;
-	        }
+            // if this is a multi-line example, set the first example to be a shared example
+            if (this.ExampleBuilders != null)
+                this.defaultMarshaller.MarshalLabel(this.ExampleBuilder.Context, SharedLabel.Instance);
         }
 
         /// <summary>
@@ -212,175 +181,21 @@ namespace VW.Serializer
         {
             if (disposing)
             {
-                if (this.DefaultNamespaceContext != null)
+                // cleanup in case CreateExample() wasn't called
+                if (this.ExampleBuilder != null)
                 {
-                    this.DefaultNamespaceContext.Dispose();
-                    this.DefaultNamespaceContext = null;
+                    this.ExampleBuilder.Dispose();
+                    this.ExampleBuilder = null;
                 }
 
-                if (this.Context != null)
+                if (this.ExampleBuilders != null)
                 {
-                    this.Context.Dispose();
-                    this.Context = null;
+                    foreach (var eb in this.ExampleBuilders)
+                        eb.Dispose();
+
+                    this.ExampleBuilders = null;
                 }
             }
-        }
-
-        private void ParseLabel()
-        {
-            // peak the first property name
-            if (!reader.Read())
-                throw new VowpalWabbitJsonException(reader.Path, "Unexpected end");
-
-            switch (reader.TokenType)
-            {
-                case JsonToken.StartObject:
-                    {
-                        // parse complex object
-                        if (!reader.Read() || reader.TokenType != JsonToken.PropertyName)
-                            throw new VowpalWabbitJsonException(reader.Path, "Expected at least a single property to determine the label object");
-
-                        var propertyName = (string)reader.Value;
-
-                        var prefixReader = new PrefixedJsonReader(this.reader,
-                            Tuple.Create(JsonToken.StartObject, (object)null),
-                            Tuple.Create(JsonToken.PropertyName, (object)propertyName));
-
-                        Type labelType;
-                        if (!labelPropertyMapping.TryGetValue(propertyName, out labelType))
-                            throw new VowpalWabbitJsonException(reader.Path, "The first property ('" + propertyName + "') must match to a property of a VowpalWabbit label type.");
-
-                        var label = (ILabel)jsonSerializer.Deserialize(prefixReader, labelType);
-
-                        this.defaultMarshaller.MarshalLabel(this.Context, label);
-                    }
-                    break;
-                case JsonToken.Integer:
-                case JsonToken.Float:
-                case JsonToken.String:
-                    {
-                        // pass label directly to VW
-                        var labelString = reader.Value.ToString();
-
-                        this.Context.ExampleBuilder.ParseLabel(labelString);
-                        // prefix with label
-                        this.Context.AppendStringExample(false, "{0}", labelString);
-                    }
-                    break;
-                default:
-                    throw new VowpalWabbitJsonException(reader.Path, "Expected label object");
-            }
-        }
-
-        /// <summary>
-        /// Expects: "1,2.2,3]" (excluding the leading [)
-        /// </summary>
-        private void ParseFeatureArray(VowpalWabbitMarshalContext context, Namespace ns)
-        {
-            ulong index = 0;
-
-            while (reader.Read())
-            {
-                switch (reader.TokenType)
-                {
-                    case JsonToken.Integer:
-                        MarshalFloatFeature(context, ns, index, (float)(long)reader.Value);
-                        break;
-                    case JsonToken.Float:
-                        MarshalFloatFeature(context, ns, index, (float)(double)reader.Value);
-                        break;
-                    case JsonToken.EndArray:
-                        return;
-                    default:
-                        throw new VowpalWabbitJsonException(reader.Path, "Unxpected token " + reader.TokenType + " while deserializing dense feature array");
-                }
-                index++;
-            }
-        }
-
-        private static void MarshalFloatFeature(VowpalWabbitMarshalContext context, Namespace ns, ulong index, float value)
-        {
-            context.NamespaceBuilder.AddFeature(ns.NamespaceHash + index, value);
-
-            if (context.StringExample != null)
-            {
-                context.AppendStringExample(
-                    false,
-                    " {0}:" + (context.VW.Settings.EnableStringFloatCompact ? "{1}" : "{1:E20}"),
-                    index,
-                    value);
-            }
-        }
-
-        /// <summary>
-        /// Expects that actual feature value.
-        /// </summary>
-        private void ParseFeature(VowpalWabbitMarshalContext context, Namespace ns, string featureName)
-        {
-            switch (reader.TokenType)
-            {
-                case JsonToken.Float:
-                    {
-                        var feature = new PreHashedFeature(this.vw, ns, featureName);
-                        this.defaultMarshaller.MarshalFeature(context, ns, feature, (double)reader.Value);
-                    }
-                    break;
-                case JsonToken.Integer:
-                    {
-                        var feature = new PreHashedFeature(this.vw, ns, featureName);
-                        this.defaultMarshaller.MarshalFeature(context, ns, feature, (long)reader.Value);
-                    }
-                    break;
-                case JsonToken.String:
-                    {
-                        var feature = new Feature(featureName);
-                        this.defaultMarshaller.MarshalFeatureStringEscape(context, ns, feature, (string)reader.Value);
-                    }
-                    break;
-                case JsonToken.Boolean:
-                    {
-                        var feature = new PreHashedFeature(this.vw, ns, featureName);
-                        this.defaultMarshaller.MarshalFeature(context, ns, feature, (bool)reader.Value);
-                    }
-                    break;
-                case JsonToken.Comment:
-                case JsonToken.Null:
-                    // probably best to ignore?
-                    break;
-                case JsonToken.StartArray:
-                    this.ParseFeatureArray(context, ns);
-                    break;
-                default:
-                    throw new VowpalWabbitJsonException(reader.Path, "Unexpected token " + reader.TokenType + " while deserializing primitive feature");
-            }
-
-        }
-
-        /// <summary>
-        /// Parses { "feature1":1, "feature2":true, .... }
-        /// </summary>
-        private void ParseNamespaceAndFeatures(VowpalWabbitMarshalContext context, string namespaceValue)
-        {
-            var ns = new Namespace(this.vw, namespaceValue);
-            this.defaultMarshaller.MarshalNamespace(context, ns, () =>
-            {
-                while (reader.Read())
-                {
-                    switch (reader.TokenType)
-                    {
-                        case JsonToken.PropertyName:
-                            var featureName = (string)reader.Value;
-
-                            if (!reader.Read())
-                                throw new VowpalWabbitJsonException(reader.Path, "Unexpected end while parsing namespace");
-
-                            this.ParseFeature(context, ns, featureName);
-                            break;
-                        case JsonToken.EndObject:
-                            return;
-                    }
-                }
-            });
         }
     }
 }
